pyproject.toml
[project]
name = "streamlit-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Muhammad Raffey", email = "muhammadraffey26@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=0.2.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.1",
]

[project.scripts]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
